ParserProduit/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg2-binary
python-multipart

pdfplumber
pytesseract
Pillow

python-dotenv
requests
